Fix for stuck toasts (#29)
Incorrect local time could cause a negative display duration. This would cause the toast to always be displayed.

It is impractical to request server time. In this commit a solution is implement solely for the display issue. Not the underlying cause.

In addition, a constant has been added for the default toast display duration.